+++ src/AppBundle/Document/Video.php
@@ -13,6 +13,7 @@ use Gedmo\Mapping\Annotation as Gedmo;
  */
 class Video
 {
+    use Timestampable;
     /**
      * @ODM\Id
      */
