+++ src/AppBundle/Resources/views/News/getSingleNews.html.twig
@@ -4,7 +4,7 @@
 
 {% block content %}
 
-<div class="col-sm-12" id="featured">
+<div class="col-sm-12" id="featured" xmlns="http://www.w3.org/1999/html">
     <div class="page-header text-muted">
         {{ news.title }}
     </div>
@@ -18,7 +18,8 @@
 
     </div>
     <div class="col-lg-12">
-        Добавив: {{ news.author.username }}
+        Добавив: {{ news.author.username }}</br>
+        {{ news.createdAt|date("m/d/Y") }}
     </div>
     <div class="col-sm-2">
         {#<a href="#" class="pull-right"><img src="{{ news.image }}" class="img-circle"></a>#}
