+++ src/AppBundle/Resources/views/index.html.twig
@@ -21,7 +21,7 @@
         <div class="col-sm-10">
 
             <h4><span class="label label-default"><a href="{{ path('app_get_single_news', {"slug" : news.slug}) }}">{{ news.title }}</a></span></h4>
-                <small class="text-muted">{{ news.text }}1 hour ago • <a href="{{ path('app_get_single_news', {"slug" : news.slug}) }}">Read More</a></small>
+                <small class="text-muted">{{ news.text }}{{ news.createdAT|date("F jS \\a\\t g:ia") }} • <a href="{{ path('app_get_single_news', {"slug" : news.slug}) }}">Read More</a></small>
 
         </div>
         <div class="col-sm-2">
