+++ src/AppBundle/Service/AddVideoService.php
@@ -23,14 +23,9 @@ class AddVideoService extends Controller
         $parsed_url = parse_url($url);
         parse_str($parsed_url['query'], $parsed_query);
         $newlink = '<iframe src="http://www.youtube.com/embed/' . $parsed_query['v'] . '" type="text/html" width="400" height="300" frameborder="0"></iframe>';
+
         $video->setLink($newlink);
-//        $json = json_decode(file_get_contents("http://gdata.youtube.com/feeds/api/videos/".$parsed_query['v'] ."?v=2&alt=jsonc"));
-
-
         $video->setSong($song);
         $this->dm->persist($video);
-
-        $this->dm->flush();
-
     }
 }
